feat(gui.py): add GUI for creating and saving worship services with hymns
This commit adds a graphical user interface (GUI) for creating and saving worship services with hymns. The GUI is built using PySimpleGUI and consists of three pages: HOME, START, and CREATE. The HOME page has two buttons, one to start creating a new service and another to exit the application. The START page allows the user to select a date for the service and create a new service. The CREATE page allows the user to add hymns to the service and save the service. The hymns are saved in a local directory specified in the settings file.